Update marketplaceApps.json
I changed static IP to true for both the Timpi Collector and Geocore, after some complains from their community about constant IP rotation. This should improve it. 

I also updated pricing for both because of the static IP, and adjusted the storage requirement for the collector node from 20GB to 1GB, as discussed with the timpi team.